# Speed up UUID creation for `Element._id`